requirements.txt
streamlit
python-dotenv
requests
beautifulsoup4
pydantic
langchain
langchain-core
langchain-openai
langchain-community
langchain-groq
langgraph
pymongo
sentence-transformers
underthesea
torch
torchvision
rich
tavily-python
